requirements.txt
google-generativeai
python-telegram-bot
flask
gunicorn
yt-dlp
